# pkg/api/openapi.yaml
openapi: "3.0.3"
info:
  title: PicoClaw Gateway API
  description: HTTP and WebSocket API for the PicoClaw AI assistant gateway.
  version: "1.0.0"
servers:
  - url: /api/v1
    description: Gateway server
paths:
  /chat:
    post:
      summary: Send a chat message (synchronous)
      operationId: chatSync
      tags: [Chat]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatRequest'
      responses:
        '200':
          description: Chat response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatResponse'
        '400':
          description: Invalid request
        '504':
          description: Timeout
  /chat/stream:
    post:
      summary: Send a chat message (SSE streaming)
      operationId: chatStream
      tags: [Chat]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatRequest'
      responses:
        '200':
          description: SSE event stream
          content:
            text/event-stream:
              schema:
                type: string
  /chat/ws:
    get:
      summary: WebSocket chat endpoint
      operationId: chatWebSocket
      tags: [Chat]
      parameters:
        - name: session_id
          in: query
          schema:
            type: string
      responses:
        '101':
          description: WebSocket upgrade
...
components:
  schemas:
    ChatRequest:
      type: object
      required: [message]
      properties:
        message:
          type: string
        session_id:
          type: string
    ChatResponse:
      type: object
      properties:
        response:
          type: string
        session_id:
          type: string
...